Renaming lmdb.hpp and rocksdb.hpp files to component.hpp to match their namespace/class.
