+++ nano/node/wallet.hpp
@@ -7,7 +7,7 @@
 #include <nano/node/openclwork.hpp>
 #include <nano/secure/common.hpp>
 #include <nano/store/component.hpp>
-#include <nano/store/lmdb/lmdb.hpp>
+#include <nano/store/lmdb/component.hpp>
 #include <nano/store/lmdb/wallet_value.hpp>
 
 #include <atomic>
