+++ nano/store/CMakeLists.txt
@@ -11,12 +11,12 @@ add_library(
   frontier.hpp
   lmdb/account.hpp
   lmdb/block.hpp
+  lmdb/component.hpp
   lmdb/confirmation_height.hpp
   lmdb/db_val.hpp
   lmdb/final_vote.hpp
   lmdb/frontier.hpp
   lmdb/iterator.hpp
-  lmdb/lmdb.hpp
   lmdb/lmdb_env.hpp
   lmdb/iterator.hpp
   lmdb/transaction_impl.hpp
@@ -35,6 +35,7 @@ add_library(
   pruned.hpp
   rocksdb/account.hpp
   rocksdb/block.hpp
+  rocksdb/component.hpp
   rocksdb/confirmation_height.hpp
   rocksdb/db_val.hpp
   rocksdb/final_vote.hpp
@@ -44,7 +45,6 @@ add_library(
   rocksdb/peer.hpp
   rocksdb/pending.hpp
   rocksdb/pruned.hpp
-  rocksdb/rocksdb.hpp
   rocksdb/iterator.hpp
   rocksdb/transaction_impl.hpp
   rocksdb/version.hpp
@@ -63,11 +63,11 @@ add_library(
   frontier.cpp
   lmdb/account.cpp
   lmdb/block.cpp
+  lmdb/component.cpp
   lmdb/confirmation_height.cpp
   lmdb/db_val.cpp
   lmdb/final_vote.cpp
   lmdb/frontier.cpp
-  lmdb/lmdb.cpp
   lmdb/lmdb_env.cpp
   lmdb/transaction.cpp
   lmdb/online_weight.cpp
@@ -82,6 +82,7 @@ add_library(
   pruned.cpp
   rocksdb/account.cpp
   rocksdb/block.cpp
+  rocksdb/component.cpp
   rocksdb/confirmation_height.cpp
   rocksdb/db_val.cpp
   rocksdb/final_vote.cpp
@@ -90,7 +91,6 @@ add_library(
   rocksdb/peer.cpp
   rocksdb/pending.cpp
   rocksdb/pruned.cpp
-  rocksdb/rocksdb.cpp
   rocksdb/transaction.cpp
   rocksdb/version.cpp
   transaction.cpp
