+++ nano/store/lmdb/component.cpp
@@ -1,8 +1,8 @@
 #include <nano/lib/stream.hpp>
 #include <nano/lib/utility.hpp>
 #include <nano/secure/ledger.hpp>
+#include <nano/store/lmdb/component.hpp>
 #include <nano/store/lmdb/iterator.hpp>
-#include <nano/store/lmdb/lmdb.hpp>
 #include <nano/store/lmdb/wallet_value.hpp>
 #include <nano/store/version.hpp>
 #include <nano/store/versioning.hpp>
