+++ nano/store/rocksdb/block.cpp
@@ -1,6 +1,6 @@
 #include <nano/secure/parallel_traversal.hpp>
 #include <nano/store/rocksdb/block.hpp>
-#include <nano/store/rocksdb/rocksdb.hpp>
+#include <nano/store/rocksdb/component.hpp>
 
 namespace nano
 {
